Refactor statistical methods
